Fixed normalisation of HTTP in rul + small fixes
Because in real world web most of the pages uses HTTP over SSL but
sometimes the links in the page are in both formats(meaning HTTPS/HTTP
since web can resolve them correctly) we would get duplicates in the
crawled pages, thus we normalized all the links to be processed as they
were HTTP. Since curl can handle it(similarly like browser does)
we represt our data purly as HTTP.

Also added reges boundary/error check on crawling base url.

And added error handling of DOWNLOAD of a page via CURL.
So if the crawling fails, it will not cause whole application to crash
but it will return empty node with new error.